refactor: extract hooks from DocsLayout
